refactoring upload manager

+++ app/Http/Controllers/API/Admin/EducatorsController.php
@@ -5,9 +5,8 @@ namespace App\Http\Controllers\API\Admin;
 
 use App\Http\Controllers\API\BaseController;
 use App\Includes\Constant;
-use App\Includes\Helper;
+use App\Includes\UploadManager;
 use App\Models\Educator;
-use App\Models\UploadTransaction;
 use Illuminate\Http\Request;
 
 
@@ -20,17 +19,10 @@ class EducatorsController extends BaseController
         $educator->last_name = $request->input("last_name");
         $educator->bio = $request->input("bio");
 
-        $uk = $request->input('upload_key');
-        $upload_transaction = UploadTransaction::where('upload_key', $uk)->first();
-        if ($upload_transaction) {
-            $uc = new UploadController();
-            $uc->moveFileToFtp($uk, tenant()->id, 1, $upload_transaction->file_type, null);
-            $educator->image = $uk;
-        }
+        $result = UploadManager::saveFile(tenant()->id, 1, $educator, 'image', false, $request->input('upload_key'));
+        if ($result == Constant::$SUCCESS) $educator->save();
 
-        $educator->save();
-
-        return $this->sendResponse(Constant::$SUCCESS,  ['educator_id' => $educator->id]);
+        return $this->sendResponse($result,  ['educator_id' => $educator->id]);
     }
 
     public function updateEducator(Request $request)
@@ -39,55 +31,37 @@ class EducatorsController extends BaseController
         $educator->first_name = $request->input("first_name");
         $educator->last_name = $request->input("last_name");
         $educator->bio = $request->input("bio");
-        $educator->save();
 
         // Check for file state
         $file_state = $request->input("file_state");
         if (!$file_state) return $this->sendResponse(Constant::$NO_FILE_STATE, null);
 
         // Update related File
-        $uc = new UploadController();
-        $uk = $request->input('upload_key');
-        $result = $uc->updateFileState($file_state, tenant()->id, 1, $educator, "image", $uk);
-
-        return $this->sendResponse($result, null);
-    }
-
-    public function editEducatorImage(Request $request)
-    {
-        $educator = Educator::find($request->input('educator_id'));
-        $action = $request->input('action');
-        $file = $request->file('file');
-
-        if ($action != Constant::$FILE_ACTION_DELETE) {
-            $size = $file->getSize() / 1024;
-            if ($size > Constant::$LOGO_SIZE_LIMIT)
-                return $this->sendResponse(
-                    Constant::$FILE_SIZE_LIMIT_EXCEEDED,
-                    ['limit' => Constant::$LOGO_SIZE_NAME_LIMIT . "kb"]
-                );
-        }
-
-        Helper::uploadFileToDisk(
-            $action,
+        $result = UploadManager::updateFileState(
+            $file_state,
+            tenant()->id,
+            1,
             $educator,
             'image',
-            'public',
-            'images/educators',
-            '.png',
-            $file
+            false,
+            $request->input('upload_key')
         );
+        if ($result == Constant::$SUCCESS) $educator->save();
 
-        return $this->sendResponse(Constant::$SUCCESS, null);
+        return $this->sendResponse($result, null);
     }
 
     public function deleteEducator(Request $request)
     {
         $educator = Educator::find($request->input('educator_id'));
-        $educator->courses()->detach();
-        $educator->delete();
 
-        return $this->sendResponse(Constant::$SUCCESS, null);
+        $result = UploadManager::deleteFile(tenant()->id, $educator->image);
+        if ($result == Constant::$SUCCESS){
+            $educator->courses()->detach();
+            $educator->delete();
+        }
+        
+        return $this->sendResponse($result, null);
     }
 
     public function fetchEducators(Request $request)
@@ -103,19 +77,5 @@ class EducatorsController extends BaseController
         })->toArray();
 
         return $this->sendResponse(Constant::$SUCCESS, $educators);
-    }
-
-    public function getImage(Request $request, $educator_id)
-    {
-        $educator = Educator::find($educator_id);
-        if ($educator && $educator->image) {
-            $path = storage_path("app\public\\" . $educator->image);
-            $path = str_replace('/', '\\', $path);
-            $headers = array(
-                'Content-Type' => 'image/png',
-            );
-
-            return response()->file($path, $headers);
-        } else return null;
     }
 }
